[TS-102] re-organize files
